refactor: delete repeated code and encapsulate canvas functional into useCanvas custom hook
