+++ src/components/Canvas/service.ts
@@ -1,40 +1,3 @@
-export function drawLine(
-  ctx: CanvasRenderingContext2D,
-  startX: number,
-  startY: number,
-  endX: number,
-  endY: number
-) {
-  ctx.beginPath();
-  ctx.moveTo(startX, startY);
-  ctx.lineTo(endX, endY);
-  ctx.stroke();
-}
-
-export const drawCircle = (
-  ctx: CanvasRenderingContext2D,
-  startX: number,
-  startY: number,
-  radius: number,
-  startAngle: number,
-  endAngle: number,
-  counterclockwise?: boolean | undefined
-) => {
-  ctx.beginPath();
-  ctx.arc(startX, startY, radius, startAngle, endAngle, counterclockwise);
-  ctx.stroke();
-};
-
-export const drawRect = (
-  ctx: CanvasRenderingContext2D,
-  startX: number,
-  startY: number,
-  width: number,
-  height: number
-) => {
-  ctx.fillRect(startX, startY, width, height);
-};
-
 export async function getImageAsBase64(url: string): Promise<string | null> {
   try {
     const response = await fetch(url);
